Fix after review Renamed to `getRoutingTable` and `withRoutingContext`
