+++ driver/src/main/java/org/neo4j/driver/internal/util/ServerVersion.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.driver.internal.cluster;
+package org.neo4j.driver.internal.util;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -44,7 +44,6 @@ public class ServerVersion
     public static final ServerVersion v3_2_0 = new ServerVersion(3, 2, 0);
     public static final ServerVersion v3_1_0 = new ServerVersion(3, 1, 0);
     public static final ServerVersion v3_0_0 = new ServerVersion(3, 0, 0);
-    public static final ServerVersion v3_2_0 = new ServerVersion(3, 2, 0);
 
     public static ServerVersion version( Driver driver )
     {
