+++ driver/src/main/java/org/neo4j/driver/v1/Config.java
@@ -75,7 +75,7 @@ public class Config
     private final int connectionTimeoutMillis;
     private final RetrySettings retrySettings;
 
-    private final Map<String, String> routingParameters;
+    private final Map<String, String> routingContext;
 
     private Config( ConfigBuilder builder)
     {
@@ -92,7 +92,7 @@ public class Config
         this.connectionTimeoutMillis = builder.connectionTimeoutMillis;
         this.retrySettings = builder.retrySettings;
 
-        this.routingParameters = builder.routingParameters;
+        this.routingContext = builder.routingContext;
     }
 
     /**
@@ -187,7 +187,7 @@ public class Config
 
     RoutingSettings routingSettings()
     {
-        return new RoutingSettings( routingFailureLimit, routingRetryDelayMillis, routingParameters );
+        return new RoutingSettings( routingFailureLimit, routingRetryDelayMillis, routingContext );
     }
 
     RetrySettings retrySettings()
@@ -210,7 +210,7 @@ public class Config
         private long routingRetryDelayMillis = TimeUnit.SECONDS.toMillis( 5 );
         private int connectionTimeoutMillis = (int) TimeUnit.SECONDS.toMillis( 5 );
         private RetrySettings retrySettings = RetrySettings.DEFAULT;
-        private Map<String,String> routingParameters = null;
+        private Map<String,String> routingContext = null;
 
         private ConfigBuilder() {}
 
@@ -480,17 +480,17 @@ public class Config
         }
 
         /**
-         * Specify routing parameters that would be passed to server in getServers Procedure call for customized
+         * Specify routing context that would be passed to server in getRoutingTable Procedure call for customized
          * routing table reply.
          * This parameter is only valid for the routing driver, a.k.a. the driver created use bolt+routing in URI
          * scheme with 3.2+ Neo4j Casual Cluster servers.
-         * @param parameters The parameters to pass to getServers Procedure
+         * @param context The routing context to pass to getRoutingTable Procedure
          * @since 1.3
          * @return this builder
          */
-        public ConfigBuilder withRoutingParameters( Map<String, String> parameters )
+        public ConfigBuilder withRoutingContext( Map<String, String> context )
         {
-            this.routingParameters = parameters;
+            this.routingContext = context;
             return this;
         }
