+++ driver/src/test/java/org/neo4j/driver/v1/integration/SummaryIT.java
@@ -43,8 +43,8 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
-import static org.neo4j.driver.internal.cluster.ServerVersion.v3_1_0;
-import static org.neo4j.driver.internal.cluster.ServerVersion.version;
+import static org.neo4j.driver.internal.util.ServerVersion.v3_1_0;
+import static org.neo4j.driver.internal.util.ServerVersion.version;
 
 public class SummaryIT
 {
